+++ src/types/decorators.ts
@@ -1,7 +1,7 @@
 import { GraphQLScalarType } from "graphql";
 import { ValidatorOptions } from "class-validator";
 
-export type TypeValue = ClassType | GraphQLScalarType | Function | object;
+export type TypeValue = ClassType | GraphQLScalarType | Function | object | symbol;
 export type ReturnTypeFuncValue = TypeValue | [TypeValue];
 
 export type TypeValueThunk = (type?: void) => TypeValue;
@@ -25,8 +25,8 @@ export interface ValidateOptions {
 export type BasicOptions = TypeOptions & DescriptionOptions;
 export type AdvancedOptions = BasicOptions & DepreciationOptions;
 
-export interface ClassType {
-  new (): any;
+export interface ClassType<T = any> {
+  new (): T;
 }
 
 export interface EnumConfig {