+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 ### Features
 - add base support for GraphQL enums using TypeScript enums
+- add support for defining GraphQL unions
 ### Fixes
 - fix not working array type notation in circular dependencies (correct thunk generation)